修复bug

+++ App/yh/c/merchant/Application.php
@@ -138,9 +138,9 @@ class Application extends HttpController {
      */
     private function clean(array $arr){
         foreach ($arr as $v) {
-            if(file_exists(ROOT.$v)){
+            if(file_exists(ROOT.$v) && !is_dir(ROOT.$v)){
                 unlink(ROOT.$v);
-            }elseif(file_exists($v)){
+            }elseif(file_exists($v) && !is_dir($v)){
                 unlink($v);
             }
         }
